requirements.txt
PyOpenGL~=3.1.6
PyOpenGL-accelerate~=3.1.6
pygame~=2.5.2
freetype-py~=2.4.0
